refactor: Streamline code structure for enhanced maintainability
