+++ siren-sdk/src/main/java/com/keyvalue/siren/androidsdk/core/elements/CircularLoader.kt
@@ -1,4 +1,4 @@
-package com.keyvalue.siren.androidsdk.helper
+package com.keyvalue.siren.androidsdk.core.elements
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
