+++ siren-sdk/src/main/java/com/keyvalue/siren/androidsdk/presenter/CorePresenter.kt
@@ -1,0 +1,4 @@
+package com.keyvalue.siren.androidsdk.presenter
+
+class CorePresenter(var userToken: String, var recipientId: String) {
+}
